Refactor BaseComplexDataType structure
Moved `BaseComplexDataType` and related utilities to `libs` directory for better modularization. Enhanced `BaseComplexDataType` class with a consistent symbol-based instance check. Added more comprehensive test cases for `isIBaseComplexDataType` function to ensure expected behavior. This change improves code organization and enhances maintainability.
